Moved monk (field population) routes to database.js
